requirements.txt
boto3==1.40.1
fastapi==0.116.1
joblib==1.5.1
matplotlib==3.10.3
mlflow==3.1.1
numpy==1.26.4
pandas==2.3.1
psycopg2-binary==2.9.10
pydantic==2.11.7
python-dotenv==1.1.1
s3fs==0.4.2
scikit-learn==1.7.0
seaborn==0.13.2
uvicorn==0.35.0
xgboost==3.0.2